+++ CHANGELOG.md
@@ -8,9 +8,14 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 ### Changed
+### Fixed
+
+## [0.2.1] - 2021-08-25
+### Added
+- New operator `match_or_null`. Will resolve to `(col = value OR col IS NULL)`.
+### Changed
 - SQL with duplicate columns now specifies the column name.
 - parameter with IN operator can receive an empty array. The condition effectively becomes false (1 = 2).
-### Fixed
 
 ## [0.2.0] - 2019-10-07
 ### Changed